Feat: Add parabolic bullet motion and animation Fix: Update teleport logic and itemData type
Implemented parabolic movement at the end of bulletPrefab.

Added new bullet sprite and animation.

Fixed bidirectional teleport issue.

Changed itemData type from int to float for precision.

+++ Assets/02.Script/Item/ItemData.cs
@@ -6,6 +6,6 @@ public class ItemDataSO : ScriptableObject
     public int ID;
     public string itemName;
     public ItemType type;
-    public int power;
+    public float power;
     public string description;
 }
